Moved `child_lab_framework.core.detection` to `child_lab_framework.task.camera.detection`
