+++ child_lab_framework/_cli/cli.py
@@ -10,9 +10,9 @@ from .._procedure import calibrate as calibration_procedure
 from .._procedure import demo_sequential
 from .._procedure import estimate_transformations as transformation_procedure
 from ..core.calibration import Calibration
-from ..core.detection import chessboard, marker
 from ..core.file import load, save
 from ..core.video import Input
+from ..task.camera.detection import chessboard, marker
 
 # NOTE: This CLI is made for development purposes. It may not be a part of the final library.
