+++ child_lab_framework/task/camera/detection/arudice.py
@@ -3,8 +3,8 @@ from typing import Self
 
 import numpy as np
 
-from .. import serialization
-from ..transformation import EuclideanTransformation
+from ....core import serialization
+from ....core.transformation import EuclideanTransformation
 from . import marker
 
 # Design idea:
